requirements.txt
ultralytics>=8.0.200
opencv-python
pillow
pandas